backend/services/algorithms/fft_algorithm.py
"""
Implementação do algoritmo FFT para compressão de imagens
"""

import os
import numpy as np
from PIL import Image
import cv2
from pathlib import Path
from typing import Dict, Any
import time

from core.config import settings
from core.logging import logger

class FFTAlgorithm:
    """Algoritmo de compressão FFT"""

    def __init__(self):
        self.name = "FFT Compression"
        self.description = "Compressão baseada em Transformada Rápida de Fourier"

    def process_image(self, image_path: str, parameters: Dict[str, Any]) -> Dict[str, Any]:
        """
        Processa uma imagem usando FFT

        Args:
            image_path: Caminho para a imagem de entrada
            parameters: Parâmetros do algoritmo (frequency_cutoff)

        Returns:
            Dicionário com resultados do processamento
        """
        start_time = time.time()

        try:
            # Carregar imagem
            image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
            if image is None:
                raise ValueError(f"Não foi possível carregar a imagem: {image_path}")

            original_shape = image.shape
            logger.info(f"Processando imagem FFT: {original_shape}")

            # Converter para float32
            image_float = image.astype(np.float32)

            # Aplicar FFT 2D
            fft = np.fft.fft2(image_float)
            fft_shifted = np.fft.fftshift(fft)

            # Criar máscara de frequência
            rows, cols = image.shape
            crow, ccol = rows // 2, cols // 2

            # Parâmetro de corte de frequência
            cutoff_percent = parameters.get('frequency_cutoff', 0.3)
            radius = int(min(crow, ccol) * cutoff_percent)

            # Criar máscara circular
            y, x = np.ogrid[:rows, :cols]
            mask = (x - ccol) ** 2 + (y - crow) ** 2 <= radius ** 2

            # Aplicar filtro (manter apenas frequências dentro do raio)
            fft_filtered = fft_shifted * mask.astype(np.float32)

            # Transformada inversa
            fft_ishift = np.fft.ifftshift(fft_filtered)
            reconstructed = np.fft.ifft2(fft_ishift)
            reconstructed = np.abs(reconstructed)

            # Normalizar para 0-255
            reconstructed = cv2.normalize(reconstructed, None, 0, 255, cv2.NORM_MINMAX)
            reconstructed_image = reconstructed.astype(np.uint8)

            # Calcular métricas
            psnr = self._calculate_psnr(image, reconstructed_image)
            ssim = self._calculate_ssim(image, reconstructed_image)
            mse = self._calculate_mse(image, reconstructed_image)

            # Calcular taxa de compressão (estimativa baseada no filtro)
            compression_ratio = 1.0 / (cutoff_percent ** 2)  # Aproximação

            # Salvar imagem processada
            output_filename = f"fft_{Path(image_path).stem}_cutoff{cutoff_percent}.jpg"
            output_path = settings.PROCESSED_DIR / output_filename
            cv2.imwrite(str(output_path), reconstructed_image)

            processing_time = time.time() - start_time

            logger.info(f"FFT processado com sucesso: PSNR={psnr:.2f}, SSIM={ssim:.3f}, Ratio={compression_ratio:.2f}")

            return {
                "success": True,
                "algorithm": "fft",
                "output_path": str(output_path),
                "output_url": f"/processed/{output_filename}",
                "metrics": {
                    "psnr": round(psnr, 2),
                    "ssim": round(ssim, 3),
                    "mse": round(mse, 4),
                    "compression_ratio": round(compression_ratio, 2),
                    "file_size": output_path.stat().st_size
                },
                "parameters_used": {
                    "frequency_cutoff": cutoff_percent,
                    "radius": radius
                },
                "processing_time": round(processing_time, 2)
            }

        except Exception as e:
            logger.error(f"Erro no processamento FFT: {str(e)}")
            return {
                "success": False,
                "algorithm": "fft",
                "error": str(e),
                "processing_time": round(time.time() - start_time, 2)
            }

    def _calculate_psnr(self, original: np.ndarray, reconstructed: np.ndarray) -> float:
        """Calcula PSNR entre duas imagens"""
        mse = self._calculate_mse(original, reconstructed)
        if mse == 0:
            return float('inf')

        max_pixel = 255.0
        psnr = 20 * np.log10(max_pixel / np.sqrt(mse))
        return psnr

    def _calculate_mse(self, original: np.ndarray, reconstructed: np.ndarray) -> float:
        """Calcula MSE entre duas imagens"""
        return np.mean((original.astype(np.float64) - reconstructed.astype(np.float64)) ** 2)

    def _calculate_ssim(self, original: np.ndarray, reconstructed: np.ndarray) -> float:
        """Calcula SSIM entre duas imagens"""
        try:
            ssim_score = cv2.quality.QualitySSIM_compute(
                cv2.cvtColor(original, cv2.COLOR_GRAY2BGR) if len(original.shape) == 2 else original,
                cv2.cvtColor(reconstructed, cv2.COLOR_GRAY2BGR) if len(reconstructed.shape) == 2 else reconstructed
            )[0]
            return ssim_score
        except:
            # Fallback simples
            return 0.8  # Valor aproximado
