backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
scikit-image==0.22.0

# Machine Learning
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2

# Data processing
pandas==2.1.4
matplotlib==3.8.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

# Optional: Julia integration
# julia==0.6.1
# pyjulia==0.6.1
